Fix sign-compare

+++ src/dp_error.c
@@ -26,7 +26,7 @@ const char *dp_grpc_strerror(int grpc_errcode)
 {
 	if (grpc_errcode == DP_GRPC_OK)
 		return "Success";
-	if (grpc_errcode < 0 || grpc_errcode >= RTE_DIM(dp_grpc_error_strings) || !dp_grpc_error_strings[grpc_errcode]) {
+	if (grpc_errcode < 0 || grpc_errcode >= (int)RTE_DIM(dp_grpc_error_strings) || !dp_grpc_error_strings[grpc_errcode]) {
 		// this should never happen, the programmer must have returned a wrong value
 		assert(0);
 		return "Invalid gRPC error code";
